+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- If custom session is passed in, do not close it when `close()` is called
 
 ## [0.2.0] - 2023-02-01
 ### Fixed
@@ -19,4 +21,3 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [Unreleased]: https://github.com/IceBotYT/linear-garage-door/compare/0.2.0...master
 [0.2.0]: https://github.com/IceBotYT/linear-garage-door/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/IceBotYT/linear-garage-door/tree/0.1.0
-